+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-10T15:38:22.518Z",
+    "2026-09-11T00:37:36.779Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {