+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3949,7 +3949,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M25ZGXQ0G5Q3EE54BWTK6PA2"),
+  id: z.string().default("01M26YC9A3E01VQ5HVPFAAWG97"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>